Udpate another test

+++ planetscale/workflows_test.go
@@ -104,7 +104,7 @@ func TestWorkflows_Create(t *testing.T) {
 	c.Assert(err, qt.IsNil)
 	c.Assert(workflow.ID, qt.Equals, wantID)
 	c.Assert(workflow.Name, qt.Equals, "shard-table")
-	c.Assert(workflow.Number, qt.Equals, 1)
+	c.Assert(workflow.Number, qt.Equals, uint64(1))
 	c.Assert(workflow.SourceKeyspace.Name, qt.Equals, "source-keyspace")
 	c.Assert(workflow.TargetKeyspace.Name, qt.Equals, "target-keyspace")
 	c.Assert(workflow.Branch.Name, qt.Equals, "branch")
